Cors

+++ Coursera/server-side/routes/promoRouter.js
@@ -5,13 +5,14 @@ const app = express();
 const promoRouter = express.Router();
 var mongoose = require("mongoose");
 const promos = require("../models/promotions");
-
+var authenticate = require("../authenticate");
+var cors = require("./cors");
 app.use(bodyParser.json());
 
 promoRouter
   .route("/")
-
-  .get((req, res, next) => {
+  .options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+  .get(cors.cors, (req, res, next) => {
     promos
       .find({})
       .then(
@@ -27,46 +28,58 @@ promoRouter
       });
   })
 
-  .post((req, res, next) => {
-    promos
-      .create(req.body)
-      .then(
-        (promos) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(promos);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => {
-        next(err);
-      });
+  .post(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+    if (authenticate.verifyAdmin(req.user)) {
+      promos
+        .create(req.body)
+        .then(
+          (promos) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(promos);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => {
+          next(err);
+        });
+    } else {
+      res.statusCode = 403;
+      res.setHeader("Content-Type", "application/text");
+      res.end("You are not authorized to perform this operation!");
+    }
   })
 
-  .put((req, res, next) => {
+  .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end("PUT operation not supported on /promos");
   })
 
-  .delete((req, res, next) => {
-    promos
-      .remove({})
-      .then(
-        (resp) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(resp);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => {
-        next(err);
-      });
+  .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+    if (authenticate.verifyAdmin(req.user)) {
+      promos
+        .remove({})
+        .then(
+          (resp) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(resp);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => {
+          next(err);
+        });
+    } else {
+      res.statusCode = 403;
+      res.setHeader("Content-Type", "application/text");
+      res.end("You are not authorized to perform this operation!");
+    }
   });
 
 promoRouter
   .route("/:promoId")
-  .get((req, res, next) => {
+  .get(cors.cors, (req, res, next) => {
     promos
       .findById(req.params.promoId)
       .then(
@@ -82,41 +95,53 @@ promoRouter
       });
   })
 
-  .post((req, res, next) => {
+  .post(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
     res.statusCode = 403;
     res.end("POST operation not supported on /promos/" + req.params.promoId);
   })
 
-  .put((req, res, next) => {
-    promos
-      .findByIdAndUpdate(req.params.promoId, { $set: req.body, new: true })
-      .then(
-        (promos) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(promos);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => {
-        next(err);
-      });
+  .put(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+    if (authenticate.verifyAdmin(req.user)) {
+      promos
+        .findByIdAndUpdate(req.params.promoId, { $set: req.body, new: true })
+        .then(
+          (promos) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(promos);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => {
+          next(err);
+        });
+    } else {
+      res.statusCode = 403;
+      res.setHeader("Content-Type", "application/text");
+      res.end("You are not authorized to perform this operation!");
+    }
   })
 
-  .delete((req, res, next) => {
-    promos
-      .findByIdAndRemove(req.params.promoId)
-      .then(
-        (resp) => {
-          res.statusCode = 200;
-          res.setHeader("Content-Type", "application/json");
-          res.json(resp);
-        },
-        (err) => next(err)
-      )
-      .catch((err) => {
-        next(err);
-      });
+  .delete(cors.corsWithOptions, authenticate.verifyUser, (req, res, next) => {
+    if (authenticate.verifyAdmin(req.user)) {
+      promos
+        .findByIdAndRemove(req.params.promoId)
+        .then(
+          (resp) => {
+            res.statusCode = 200;
+            res.setHeader("Content-Type", "application/json");
+            res.json(resp);
+          },
+          (err) => next(err)
+        )
+        .catch((err) => {
+          next(err);
+        });
+    } else {
+      res.statusCode = 403;
+      res.setHeader("Content-Type", "application/text");
+      res.end("You are not authorized to perform this operation!");
+    }
   });
 
 module.exports = promoRouter;
